refactor: modify name of common constant and type
